add more restrictive right to the guest user

+++ src/users/users.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, Inject, NotFoundException } from '@nestjs/common';
+import { Injectable, Inject, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { User } from '../entities/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserRepository } from './user.repository';
 import { GetUsersFilterDto } from './dto/get-users-filter.dto';
+import { UserRole } from './user-role.enum';
 
 @Injectable()
 export class UsersService {
@@ -16,11 +17,18 @@ export class UsersService {
         return this.userRepository.getUsers(filterDto, user);
     }
 
-    async getUserById(id: number) {
+    async getUserById(id: number, user: User) {
+        if (user.role !== UserRole.ADMIN && user.id !== id) {
+            throw new UnauthorizedException('unsuccessful right');
+        }
         const found = await this.userRepository.findOne({ where: {id}, relations: ['guestInfo']});
         if (!found) {
             throw new NotFoundException('User with this ID not found');
         }
+
+        delete user.password;
+        delete user.salt;
+
         return found;
     }
 
@@ -28,7 +36,10 @@ export class UsersService {
         return this.userRepository.createUser(createUserDto, user);
     }
 
-    async deleteUser(id: number): Promise<void> {
+    async deleteUser(id: number, user: User): Promise<void> {
+        if (!this.userRepository.adminVerification(user)) {
+            return;
+        }
         const result = await this.userRepository.delete({id});
         if (result.affected === 0) {
             throw new NotFoundException('User with this ID not found');
